[M920-014, MA25-036] Simplify translation of loop (in)variants
Now that loop invariants and loop variants can only be grouped
together when they apply to the same loop, simplify their translation
into Why3. In particular, the first loop (in)variant determines the
position of the start of the generated loop in Why3, which increases
the precision of results when a Loop_Variant appears alone in a loop.

* why/gnat2why-expr-loops.adb
(Get_Loop_Invariant): Simplify getting (in)variants from the loop.
(Transform_Loop_Body_Statements): Simplify treatment as loop variants should
  not occur anymore.
(Transform_Loop_Statement): Propagate effects of simplification of
  Transform_Loop_Body_Statements.