[MB21-048] flow - local constants
Add a hack to deal with local constants: they are now treated as if they
were variables, and ineffective-initialisation messages are suppressed
for them.

The error messages are suppressed as the compiler generates some constants
which are used in type bounds - and thus they disappear as these still
do not have flow bounds.