[M710-053] Issue error on types with partial default initialization
Types with partial default initialization are forbidden by rule 3.8(1)
of SPARK RM. This implements this check. Types defined in axiomatized
units and private types whose completion is not in SPARK_Mode=On do not
raise this error. Instead they are considered to be not initialized by
default in GNATprove.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Issue violation on partially initialized type.
* gnat2why/spark/spark_util.adb|s
(Default_Initialization): Move here the corresponding function from
  Sem_Util, correct issues with private base types and protected
  types, handle specially case of types defined in axiomatized
  units and private types.