MB20-049 Support for Old in contract cases
The new semantics of Old in consequence expressions inside contract cases
is that the corresponding prefix expressions are evaluated on entry to
the subprogram only when the corresponding contract case is enabled (that
is, the corresponding guard expression evaluates to true). This requires
adapting the Why3 program, so that generation of VCs for possible run-time
errors in these prefixes is guarded by a test that the corresponding
guard expression is true.

We use this opportunity to factor out the computation of the value of the
guard expressions, so that it is not repeated as was done previously for
entry checks (disjointness and completeness of cases) and exit checks
(checking that contract cases consequences are respected). So the guard
expression is now computed only once, and the corresponding temporary
variable name is used whenever this value is required.

* gnat2why/why/gnat2why-expr.adb|s
(Bind_From_Mapping_In_Expr): Add parameters to pass in info on guards,
  when mapping value of prefixes to Old. Use guard info to evaluate Old
  prefixes under correct condition, when inside contract case.

* gnat2why/why/gnat2why-subprograms.adb
(Compute_Contract_Cases_Guard_Map): New procedure to factor out the
  computation of info related to guards in contract cases.
(Compute_Contract_Cases_Entry_Checks, Compute_Contract_Cases_Exit_Checks):
  use Guard_Map computed once, do not bind values of the temps for guards
  as these will be bound outside these functions.
(Generate_VCs_For_Subprogram): Call new procedure to get guards, and
  restructure code to bind guard temporaries in right context.

* testsuite/gnatprove/tests/MB20-049__old_in_contract_cases/test.opt
Removed, as test passes now