MC19-006 remove ALI_Closure mode for analysis with gnat2why
gnat2why in analysis mode (flow and Why generation) was run with
ALI_Closure mode, so run on more units than analysis was requested for.
This turns out to be unnecessary, this patch disables this.

As a consequence, gnat2why does not need to know the set of "analyzed
files", it will only be called on files to be analyzed.

gnat2why.cgpr
  (Dependency_Kind): changed back to ALI_File
gnat2why_args.adb
  (Set)
  (Interpret_Token)
    remove code that deals with Analyze_File argument
spark_util.adb
  (Analysis_Requested) simplify check whether the subprogram should be
    analyzed
gnatprove.adb
  (Translate_To_Why) we can now add switch "-u" to this invocation of
  gprbuild, when gnatprove was called with -u, so that no other files
  are analyzed.