MA08-009 flow - regression on initialization warnings
We now issue warnings about arrays that might be uninitialized
on the node that actually uses them (if such a node exists)
instead of issuing them on the node that causes them to be
potentially uninitialized.

Updated 10 tests that were affected by this change and added
a small test specifically for this TN.

* flow-analysis.adb
(Find_Use_Of_Uninitialized_Variables): Added an extra parameter
on nested function Might_Be_Defined_In_Other_Path to capture
where we want the error/warning message to be issued.