N107-040 multiple temp vars for one entity
GNATprove does not fail anymore when Binding_For_Temp is called
several times on the same expression that does not need a temp
variable.

* why-gen-expr.adb
(New_Temp_For_Expr): Do not store in the Temp_Names_Map map
expressions that do not need a temp variable.
(Binding_For_Temp): Ignore expressions that are not contained in the
Temp_Names_Map map.