M110-032 derived types of private types with external axioms
Properly handle subtypes and derived types of private types declared in
packages with External_Axiomatization.

* gnat2why-expr.adb
(One_Level_Access): If the access is to a discriminant of a private type
declared in a package with External_Axiomatization, introduce a
conversion to the base type before using the access function.

* gnat2why-types.adb
(Translate_Type): Fix translation for private types based on packages
with external axioms.

* why-gen-expr.adb
(Insert_Private_Conversion): New function to handle conversion
between subtypes of a private type based on a package with external
axioms.
(Insert_Checked_Conversion): Call Insert_Private_Conversion.

* why-gen-names.adb
(Conversion_Name): Handle the case of private types based on a
package with external axioms.

* why-gen-records.adb
(Insert_Subtype_Discriminant_Check): Handle the case of private types
based on a package with external axioms.

* why_inter.adb
(Is_Ext_Axioms_Conversion): New function to decide if the function
Insert_Private_Conversion must be called for a conversion.
(To_Why_Id): Fix case of discriminant of a private type declared in a
package with External_Axiomatization

* ada__model.mlw
(Unconstrained_Private_Type): New module for an unconstrained
subtype or derived type of a private type based on a package with
external axioms.
(Constrained_Private_Type): New module for a constrained
subtype or derived type of a private type based on a package with
external axioms.