N116-016 Precise analysis of bounds for floating-points
Add axioms to bound the difference between a real number and its
rounding to floating-point (both single and double precision).
This allows proving tight bounds on the result of floating-point
operations, when bounds on the arguments are known.

* share/spark/theories/_gnatprove_standard.mlw
(Floating): Add axioms to bound the result of round_single and
  round_double.

* testsuite/gnatprove/tests/N116-016__floating_point/precise.adb
* testsuite/gnatprove/tests/N116-016__floating_point/test.out
* testsuite/gnatprove/tests/N116-016__floating_point/test.py
* testsuite/gnatprove/tests/N116-016__floating_point/test_18.adb
* testsuite/gnatprove/tests/N116-016__floating_point/test_float.adb
New test

* testsuite/gnatprove/tests/K719-019__flight_mgr/test.py
* testsuite/gnatprove/tests/MC13-026__float_div/test.py
* testsuite/gnatprove/tests/MC17-036__gnatVa/test.py
Increase step limit to get same proof results as before.

* testsuite/gnatprove/tests/M809-005__float_basics/test.out
* testsuite/gnatprove/tests/openETCS_array/test.out
Accept a few less proved checks, instead of increasing step limit,
as these tests are already long enough.