MB27-015 flow - implement SPARK RM 4.4(2)
Fix a bug in the sanity checking for SRM 4.4(2) - we need to consider all
flow IDs, not just direct mappings.

Fix a follow-on issue (discussed in N116-011) where we raise un-fixable
errors on compiler-generated types. We avoid this by not sanity checking
anything that does not come from source.

Updated/fixed some tests that violated SRM 4.4(2), which we now detect.