M109-040 flow - elaboration of nested packages
Change how we analyze nested packages.

First I have introduced the notion of a variable being initialized at
specification. So in the below, X is initialized at specification, but
Y is not:

   package P with Initializes => (X, Y) is
      X : Integer := 12;
      Y : Integer;
   end P;


When we encounter a nested package specification we simply process the
visible part as if it was part of the analyzed subprogram (and also add
any abstract state variables introduced by it). So the following code:

   X : Integer;
   package P is
      Y : Integer := Foobar;
   end P;

Is simply analyzed as:

   X : Integer;
   Y : Integer := Foobar;

Note that the initializes aspect is not considered at this point.


When we encounter a nested package body, we apply the initializes aspect;
flagging any variables mentioned by it as initialized and introducing
any flow dependencies mentioned by it.