N122-016 fix incremental analysis
Previous "fix" which added "-s" to all gprbuild calls was incorrect. We
now only add -s to the call in phase 2. See the added comment for
details. To avoid rerunning gnat2why all the time because of switch
"-gnates=" changing, we use the hash of file contents for this file.

* gnat2why-args.adb
(Interpret_Token): Robustify code by accepting an empty line
(Set): Change way of creating the file and the file name

* gnatprove.adb
add comment about incremental analysis