N202-004 N206-014 flow - carry flow information through attributes
These are slight regressions introduced by N202-004 that are to
be fixed in N206-014. The problem here is that the itype for an
unconstrained array, such as a string, will stop the variable from
being flagged as unused. It should be noted that it is still flagged
as an ineffective import.