N206-017 flow - bug when calling subprogram with discr out param/global
Fix boundary issue where a subprogram with an unconstrained parameter
is called with a constrained parameter.

Fix the crash related to calling a subprogram with an unconstrained
parameter.

Update comments and tests, confirming the behaviour of calling subprograms
with unconstrained parameters or discriminated parameters is now OK.

Enable test for N202-004.