N207-018 Tidy up segway example for discussion
Tidy up this example for discussion and agreement of the standard
required for the release.

In this case

1) Comply with -gnaty

2) Add an overall comment on the purpose of the example. Note this appear
in segway.adb for now, since this file received the editor focus in GPS
when the use picks the example of the Help menu.  This depends on the
order of files opened in spark.py, which is now set in spark.py in
GPS 6.0.1, so too late to change for rel 1.

3) Add an Initialize procedure, which is a rather obvious ommission
in a state-machine package.