LC17-035 proof support for multidim 'Update
Added support in gnat2why for multi-dimensional 'Update aggregates. These
aggregates, for example (I, J, K) of

   An_Array'Update((I, J, K) => New_Val)

need special treatment when recursively translating the choices to be
parameters of the logic function (Traverse_Value_At_Index in
Generate_Logic_Function), and when producing the call to the logic function
(Select_These_Choices and Complete_Translation).

The axiom of the logic function for 'Update aggregates is now generated
with a conjunction of comparisons, one for each dimension, when selecting
the choices.

Get_Range_Check_Info now deals with ranges and choices of multi-dimensional
'Update aggregates. This makes sure checks are generated (from
Transform_Expr from Complete_Translation from Transform_Aggregate), and
relies on the Do_Range_Check_Flag set by the compiler.

Now removed the checks previously introduced by Transform_Aggregate by
modifying Get_Aggregate_Elements to only collect Index_Values for choices
of normal aggregates. Choices of normal aggregates are not used as
parameters to the logic function and therefore will not get the checks
inserted by a call to Transform_Expr in Complete_Translation.

 * spark_definition.adb
   turned on processing of multidim 'Updates

 * gnat2why/why/gnat2why-expr.adb
   modified Get_Aggregate_Elements, Complete_Translation,
   Select_These_Choices, Generate_Logic_Function in Transform_Aggregate

 * gnat2why/why/why-gen-expr.adb
   modified Get_Range_Check_Info

Testsuite:

 * turned on existing multidim tests in LC17-035__update_attribute_multidim
   and RM_Examples

 * added tests for checks:
   LC17-035__update_attribute_multidim/update_checks.ad*

 * added tests for correctness of the logic function/axiom:
   LC17-035__update_attribute_multidim/update_logic_fn.ad*

 * reviewed reference results for all, sample reviews of the generated Why
   as well.