LC17-035 Fix steps to get consistent results on all platforms
After commit on LC17-035 for multidim array Update, test
update_examples.adb in RM_Examples gets a different for the proof
of postcondition on line 38 on Windows compared to other platforms:
proved on Windows, not proved on others. This is because most runs
of gnatprove for RM_Examples are done with the default timeout of
1s instead of a fixed steps value. Correct value for
update_examples.adb, will open TN for others.