MC02-014 operators as actual of packages with external axioms
* spark_definition.adb
(Mark_Package_Entity): Do not issue error messages for operators as
actual of packages with external axioms anymore.

* gnat2why-expr.adb
(Transform_Concatenation): Change signature so it takes Why
expressions for operands and not the Ada node.
(Transform_Array_Equality): Change signature so it takes Why
expressions for operands and not the Ada node.
(Transform_Array_Comparison): Change signature so it takes Why
expressions for operands and not the Ada node.
(Transform_Array_Logical_Op): Change signature so it takes Why
expressions for operands and not the Ada node.
(Transform_Array_Negation): Change signature so it takes Why
expressions for operands and not the Ada node.
(New_Op_Expr): New function to create an operation that takes Why
expressions for the logical operands.
(Transform_Expr): Call New_Op_Expr for operators.

* gnat2why-external_axioms.adb
(Parse_Parameters): Support operators as actual parameters using
New_Op_Expr.