N309-004 Adopt same rules for Old in Contract_Cases as in Post
The frontend-enforced rules for Old appearing in Contract_Cases are
now the same as when Old appears in Post: forbidden by default, but
the user can use the GNAT pragma Unevaluated_Use_Of_Old to make it
a warning or to accept it.

* docs/lrm/source/subprograms.rst
Remove wording that made a different between Post and Contract_Cases.

* testsuite/gnatprove/tests/L316-044__subp/database.adb
* testsuite/gnatprove/tests/L316-044__subp/database.ads
* testsuite/gnatprove/tests/L321-029__limitline/database.ads
* testsuite/gnatprove/tests/database/database.ads
Adapt tests.