N317-010 Add marking mechanism in testsuite for automating review or results
Given the output from flow analysis and/or proof, check that all marks
mentioned in source files have a matching expected result, where source
files are taken to be the *.ad? files in the current directory.

Marks are any strings in the source that have the format
    @TAG:RESULT
where both TAG and RESULT are alphanumeric strings without space, possibly
with underscores. A tag denotes a line some result is expected (typically
this marker will be put in comments).

TAG is either:
- a check (RANGE_CHECK, DIVIDE_CHECK, etc), or
- a flow message (UNINIT, DEPENDS, etc).

The complete list of tags is given by functions is_flow_tag and
is_proof_tag.

RESULT is either
- PASS/FAIL for checks, or
- ERROR/WARN for flow messages.

Case does not matter for the tag or result, although UPPERCASE is better in
source code to easily locate the marks visually.

* testsuite/gnatprove/lib/python/test_support.py
(check_marks): New function issues an error if source mark not matched in
  results.
(gnatprove_): Call check_marks.
* testsuite/gnatprove/tests/K519-031__subarray/subar.adb
* testsuite/gnatprove/tests/K816-021__twodim/twodim.adb
* testsuite/gnatprove/tests/KA17-027__obj_renaming/main.adb
* testsuite/gnatprove/tests/KA20-075__constant/const.adb
* testsuite/gnatprove/tests/KB27-003__color/p.ads
* testsuite/gnatprove/tests/L531-021__discr/basic.adb
* testsuite/gnatprove/tests/MC06-015__contracts_on_intrinsics/p.adb
* testsuite/gnatprove/tests/N108-022__ada_real_time/rt.adb
* testsuite/gnatprove/tests/TU__nonreturning_procedures/nr.adb
* testsuite/gnatprove/tests/fibonacci/fib.adb
* testsuite/gnatprove/tests/fibonacci/fib.ads
Update source to add tags where checks should fail, and one place where it
is expected to pass.