N325-022 fix translation of volatile objects
Previously, the code was expecting translation of volatiles in the
"program" domain only. This is too restrictive, however, and there are
cases where gnatprove decides to translate something using the domain
term. This patch allows translation of volatiles in the domain term.
This is correct because of the general invariant that any actual Ada
code will be translated in program domain at least once (e.g. for
checks).

* gnat2why-expr.adb
(Transform_Identifier): allow translation of volatiles in domain term