N414-024 Register declared objects which are not mentioned in expr
In some cases, an object still needs to be registered although
it does not appear in an expression. This is the case for example
for a formal generic constant parameter which is simplified out in
the generic instance. It may still appear in the effects for the
instance, and as such should be registered.

* gnat2why/spark/spark_rewrite.adb
(Rewrite_Compilation_Unit): Add special case for object declarations.