N409-033 Issue error on wrong access to component
In some cases, an access to a component that does not exist in the static
type of the object is not detected by the frontend. Issue an error in
gnat2why in those cases.

* gnat2why/spark/spark_definition.adb
(Mark): Issue error on selected component when static type does not
  allow it.

* gnat2why/spark/spark_util.adb|s
(Search_Component_By_Name): Return Empty when component not found, instead
  of raising exception.

* gnat2why/why/gnat2why-expr.adb
(Transform_Record_Component_Associations): Add Assert to guard against
  ill-formed AST.