N430-037 flow - limited view bug
Procedure Get_Globals used to return the limited view of a variable.
However, when generating flow graphs we operate on the full view of
the variable, so we used to get two different flow ids for the same
variable.

* flow_utility.adb
(Get_Globals): Edited nested procedure Process so that it adds the
  Non_Limited_View of the global variable (if available).

* testsuite
(N430-037__ipstack_buffer): Enabled test and updated expected output.