N404-022 Support subprograms with mutually recursive contracts
During the translation of a subprogram, generate the declaration of the Why
program function in the theory completion to support mutually recursive
contracts.

Add a third iteration over the list of entities during the translation to
generate subprograms' completions.

* why-gen-binders.ads/b
(Item_Enum): Add a new kind for functions which should be translated
differently in programs and in annotations.
(Item_Type): Extend Item_Type to the new kind.
(To_Binder_Array): Fail on the new kind as functions are not usually stored
in binder arrays.

* gnat2why-decls.adb
(Translate_Variable): Fail on new kind of Item_Type.

* gnat2why-driver.adb
(Complete_Subprogram): New local procedure to generate subprograms'
completions.
(Translate_CUnit): Call Complete_Subprogram to do the additional iteration.
(Translate_Entity): Do not generate an empty axiom theory for subprograms.

* gnat2why-expr.adb
(Transform_Identifier): Handle new kind of Item_Type.
(Transform_Quantified_Expression): Use Transform_Identifier to get the id
of functions Has_Element and Element of types with Iterable.
(Transform_Statement_Or_Declaration): Use the table of symbols to retrieve
names of procedures.

* gnat2why-external_axioms.adb
(Register_Decls): Unlike in usual cases, program functions of subprograms
declared in external axioms should be found in the main module.

* gnat2why-subprograms.adb
(Add_Dependencies_For_Effects): New procedure that generates imports for
effects of a subprogram.
(Compute_Effects): Do not generates imports for effects of a subprogram any
more.
(Generate_Subprogram_Program_Fun): New local procedure to generate a Why
program function for a subprogram.
(Generate_Axiom_For_Post): New local procedure to generate an axiom for the
contract of a subprogram.
(Generate_Subprogram_Completion): New procedure to generate a Why program
function and an axiom for the contract of a subprogram.
(Translate_Expression_Function_Body): Also generate a Why program function
and an axiom for the contract of the expression function.
(Translate_Subprogram_Spec): Only generate a logic function for Ada
functions.

* why-gen-names.ads
(Post_Axiom): New name for the axiom for the contract of a subprogram.

* why-inter.adb
(Is_Relevant_Node_For_Imports): Ignore entities of Ekind E_Void.
(To_Why_Id): In programs, the translation of a subprogram entity should be
searched in the axiom module.