N520-008 Allow objects/parameters declared "aliased" in SPARK
The mere declaration of an aliased object or parameter should not be
considered as a SPARK violation. It is the aliasing introduced by a
conversion of an address to a pointer which should be rejected.

* docs/lrm/source/declarations-and-types.rst
Update LRM to allow use of keyword "aliased"

* gnat2why/spark/spark_definition.adb
(Mark...): Do not reject uses of "aliased"

* testsuite/gnatprove/lib/python/test_support.py
* testsuite/gnatprove/tests/M326-011__trace/test.out
Sort traces in output.

* testsuite/gnatprove/tests/M930-001__flow_support_proof_in/test.out
*  testsuite/gnatprove/tests/M930-001__flow_support_proof_in/test.py
Do not use --ide-progress-bar, useless here, which does not guarantee
proper sorting of output.

* testsuite/gnatprove/tests/N127-028__aliased_formals/test.out
* testsuite/gnatprove/tests/TU__access_attribute/test.out
Update output.