N512-043 Introduce and allow Volatile_Components
Introduces definitions of "effectively volatile type" and
"effectively volatile object" to include types and objects
that are [of] array types that have Volatile_Components
specified.

Adjust legality rules to use these terms so that a Volatile_Components
array can be used in same contexts as a volatile type.

Remove use of "volatile variable" throughout and use "volatile
object" consistently.