N423-027 improve message printing by gnat2why
JSON output is now removed, instead a message id is appended to the
message, in IDE mode. Generation of the JSON files .flow and .proof is
unchanged, with the exception that the message text is now not
printed there, but the message id.

This commit also contains some refactoring and removes some complexity
of the previous implementation

* flow_error_messages.adb
(Compute_Message) new function to compute the message text
(Compute_Sloc) new function to compute the message sloc
(Add_Json_Msg) new function to save the JSON only
(Print_Regular_Msg) new function to print the compiler msg only
(Warning_Is_Suppressed) new function to check whether the message is
suppressed