N416-037 Procedure calls cannot modify bounds of unconstrained arrays
Give in out and out parameters of unconstrained array types in split
form to procedures in Why, so that their bounds are known to be
preserved.

Also refactor a bit to use the same mechanism to retrieve the expected
form of the Why parameters of a subprogram during the subprogram's
declaration and during calls' translation.

* gnat2why-expr.adb
(Needs_Temporary_Ref): Provide the expected why type for the formal
parameter.
(Compute_Call_Args): Use Compute_Subprogram_Parameters to get the
expected type and form of the subprogram arguments. Support array
arguments in a split form.
(Insert_Ref_Context): Use Compute_Subprogram_Parameters to get the
expected type and form of the subprogram arguments. Support array
arguments in a split form.

* gnat2why-subprograms.adb
(Add_Dependencies_For_Effects): Expects a theory instead of a File.
(Compute_Raw_Binders): Use split form for out and in out unconstrained
array parameters of a procedure.
(Compute_Subprogram_Parameters): New function to create binders for a
subprogram. It computes raw binders when in programs and logic
binders otherwise.
(Compute_Binders): Use Compute_Subprogram_Parameters to get the
expected type and form of the subprogram arguments.
(Add_Logic_Binders): Renaming of Compute_Logic_Binders that takes
the raw binders as an argument.

* why-gen-arrays.adb
(Array_Convert_From_Base): New version that retrieves the bounds from
an other array.