N617-015 fix crash in spark_report
spark_report expects a "message" field when the "rule" field is set to
"pragma_warning". However, in the case of a proved message within the
scope of a pragma Warnings(Off), one could have the rule field set to
"pragma_warning", but no message field present. This is now fixed.

* flow_error_messages.adb:
(Add_JSON_Msg) only set msg kind to pragma_warning when we actually have
  a warning