N522-013 Compile why3 library for Coq at run-time
* gnatprove/configuration.adb
(Prepare_Prover_Lib): New function, intended to compile why3 libraries
when necessary for interactive provers but for now only supports Coq.
The function looks for the needed binaries, if not present, it copies
the source files shipped with SPARK and compiles them in the gnatprove
directory.

(Read_Command_Line): Call to Prepare_Prover_Lib.

* why3
Added variable to modify the default location of why3's prover libraries.