N701-012 fix incorrect range check
Gnat2why correctly inserts up to two checks on passed parameters of a
subprogram call: one for fetch, one for store. However, sometimes the
check for the "store" case was not done against the correct type. Fixed
now.

* spark_util.adb
(Get_Formal_Type_Of_Actual) renamed, now returns the formal instead of
  its type
* why-gen-expr.adb
(Get_Range_Check_Info) look at the mode of the formal to determine
better the check case in which we are in (fetch/store)