N710-015 Missing detection of call to non-SPARK subprogram
The detection that a subprogram marked SPARK_Mode=>Off is called in
SPARK code was missing, leading to a crash later on when trying to
retrieve a computed effect for that subprogram. Now fixed.

* gnat2why/spark/spark_definition.adb
(Mark_Call): Detect violation of SPARK when calling non-SPARK subprogram.

* gnat2why/why/gnat2why-util.adb
Style

* testsuite/gnatprove/tests/MC12-026__overlayed_objects/p.adb
* testsuite/gnatprove/tests/MC12-026__overlayed_objects/test.out
Fix test to be legal.

* testsuite/gnatprove/tests/N710-015__spark_off/test.out
* testsuite/gnatprove/tests/N710-015__spark_off/spark_off.adb
* testsuite/gnatprove/tests/N710-015__spark_off/spark_off.ads
* testsuite/gnatprove/tests/N710-015__spark_off/test.py
New test.