N611-014 Remove common Main field in items of the symbol table
Split record do not have a main part that can be assumed to hold
all the information about the object. Dispatch the common Main
field into each variant part to prepare for the addition of a new
kind of item for them.

* why-gen-binder.ads
(Item_Type): Dispatch the common Main field into each variant
part of the record type.

* gnat2why-decls.adb
(Mk_Item_Of_Entity): Create the identifier that should be used in
the binder instead of taking it as input so that the case analysis
for the construction of the item is only done once.