N709-001 Fix references to global view of params for inherited contracts
The generated Why3 code refers to a global view of parameters when it
comes to generating VCs for a given subprogram. Inherited contracts (pre-
and postconditions) should refer to the global view of the parameters on
the current subprogram instead of the parameters of the overridden
subprogram. This is now fixed by adding the proper redirection in the
symbol table.

* gnat2why/spark/spark_util.adb|s
* gnat2why/why/gnat2why-util.adb
Adapt to version of Find_Contracts which takes Classwide or Inherited but
not both.

* gnat2why/why/gnat2why-driver.adb
(Generate_VCs): Push correct mapping for parameters of overridden
  subprograms corresponding to the inherited contracts, if any.

* gnat2why/why/gnat2why-subprograms.adb|s
(Update_Symbol_Table_For_Inherited_Contracts): New procedure sets up
  a redirection for parameters of overridden subprogram to current
  parameters.

* testsuite/gnatprove/tests/N709-001__contracts/test.out
Update improved output.