N709-001 Check AoRTE and LSP for classwide post in an empty context
The check that the classwide post is stronger than the inherited one,
and that its evaluation cannot raise a run-time error, should be performed
in an empty context, in particular not assuming the classwide pre.
Otherwise, such verification are not composable.

* gnat2why/why/gnat2why-subprograms.adb|s
(Generate_VCs_For_LSP): Do not assume Pre'Class (or inherited one) to
  check AoRTE of Post'Class and that Post'Class is stronger than inherited
  one.

* testsuite/gnatprove/tests/N709-001__contracts/test.out
Update output