N502-021 Add support for attribute Subtype'Size
Attribute Size may not always be computed statically in the frontend,
for composite types, in which case it should be translated as a correct
application of an uninterpreted function. This is only supported currently
when the prefix is a subtype, not yet when the prefix is an object.

* gnat2why/flow/flow-control_flow_graph.adb
(Pragma_Relevant_To_Flow): Use same groups for pragmas as in
  SPARK_Definition.

* gnat2why/spark/spark_definition.adb
(Mark_Pragma): Raise exception on pragmas that should not reach here.

* gnat2why/why/gnat2why-expr.adb
(Transform_Attribute): Translate attribute Size.
(Transform_Pragma): Use same groups for pragmas as in SPARK_Definition.

* gnat2why/why/why-gen-expr.adb|s
(New_Attribute_Expr, To_String): Add case for Attribute_Size.

* gnat2why/why/why-gen-names.adb|s
(Why_Name_Enum): Add name for attribute Size.

* gnat2why/why/why-gen-records.adb
(Declare_Attributes): New function to group definition of attributes
  Size and Tag.

* share/spark/theories/ada__model.mlw
Add logic function attr__size for the Size attributes to constrained
array theories.