Minor correction to issue proper error on implicit dereference
An implicit dereference may occur in a selected component. Previously,
this caused a strange error to be issued about a missing component, as
the component was not found in the type of the prefix. Now issue an
appropriate error message about implicit dereference not being supported
in SPARK.

* gnat2why/spark/spark_definition.adb
(Mark): Detect implicit dereference in selected component.

* gnat2why/spark/spark_util.ads
(Has_Access_Type): New query function.