N806-021 Complete proof of railway example
This version is now completely proved by a combination of provers, with
--timeout=20 and --proof=per_path.

The remaining unproved assertion was due to an overlooked detail in the
specification: the previous tracks are not only those that follow the
track where the tail of the train is, but *also* those that follow the
track where the head of the train is, when the train is between two
tracks! Now corrected.