N812-022 fix missing error message for non-SPARK feature
An error message was missing for a non-SPARK feature in code marked
SPARK_Mode(On). Another error message was issued, but in another unit,
which makes it very difficult to see the problem. This patch makes both
messages appear, by changing the order of marking

* spark_definition.adb
(Mark) first look at the node, then look at the Etype, this gets us
useful message