N818-021 Sending source information to Why3
* gnat2why/why/why-gen-expr.adb
(New_Comment_Label): New function creating a comment label for a given VC
with the physical source line that generated the VC. Another line
containing a marker to the corresponding column and the VC reason is also
embedded in the comment to be displayed just below the source line.
The function uses the VC location label previously generated to avoid
computing a different location than the one reported in GNATprove messages.

(New_VC_Labels): Call to New_Comment_Label.