N916-037 Restructure UG chapter on SPARK 2014 language
Move hard-to-find section on contextual analysis to a new section
in the chapter on GNATprove, about how to write subprogram contracts.
Add a subsection of this new section on generation of data and
information contracts, differentiating the precise generation for
SPARK subprogram from the coarse generation for non-SPARK subprograms.
Add examples.

Fix the Makefile for GNATprove-by-Example code snippets, so that illegal
examples do not prevent generation of error/warnings for legal examples.