N404-009 removed bug in processing of <> in aggregates
Removed assertion (Values.Length /= 0) to cope with cases where
box is present in all the associations of an array
aggregate. Currently, box is only modelled with the default true
predicate in the axiom for an array aggregate. Safe to remove the
assertion because both Generate_Logic_Function and
Complete_Translation traverse and analyse the associations,
rather than the Values.

Test update.