NA14-004 Replay facility in GNATprove always issues warning/check messages
When a unit or the units on which it depends have not changed, GNATprove
detects it and avoids re-analyzing the unit. This caused some warnings to
be issued only the first time a unit was analyzed. Now, a new replay
facility is used in GNATprove to always reissue warning and check messages
when a unit is considered, even when it does not require re-analysis.

* gnatprove/gnatprove.adb
Update top-level description in comments.
(Translate_To_Why): Rename subprogram into Flow_Analysis_And_Proof.
  Use switch --complete-output of gprbuild to replay messages on stdout
  and stderr when output is up-to-date.
* testsuite/gnatprove/tests/NA14-004__replay
New test for this feature.