NB04-022 flow - classwide globals
Implement the first step to support classwide globals.

Get_Globals has a new parameter to obtain the classwide or specific
globals. For proof we set this to the safe default "true", for flow
we are more precise right now.

Classwide globals are assumed to be equal to specific globals right
now.

Check all (classwide) global contracts for primitive operations on
tagged types that they conform to SRM 6.1.6.

Updated tests.