NA22-017 fix/improve pragma annotate for errors in spec
pragma Annotate(Gnatprove) didn't work for errors in SPARK aspects,
because the pragma would only apply to the very last declaration
preceding it, which would be one of the aspects rewritten as pragma.
This is now fixed, by considering a declaration always with the
following declarations which are not from source. So the subprogram
declaration will also include the aspects rewritten into pragmas, and
the pragma Annotate will apply to all of them.

Also, we have changed the meaning of a pragma Annotate after a body so
that it applies both to the spec and body. This allows hiding pragma
Annotate for the spec which follow from implementation details in the
body.

* spark_definition.adb
(Mark_Stmt_Or_Decl_List) differentiate between parent node or preceding
  node when calling Mark_Pragma_Annotate
* gnat2why-annotate.adb
(Insert_Annotate_Range) two new functions for lists and for slocs; also,
  special case for N_Subprogram_Body which should take care of both a
  bug in Sloc_Range and the changed semantics mentioned above
(Insert_With_Next) new function similar to Insert_With_Next which also
  inserts following nodes until hitting a source node
(Mark_Pragma_Annotate) call Insert_Annotate_Range or Insert_With_Next as
appropriate