N317-021 Add Why3 theory for formal indefinite vectors
The difference between the definite and indefinite version of formal
containers is only in the type of elements that they can contain, but
not their API which is exactly the same. As there is currently no
way to share a theory for two different Ada units, we copy/rename the
theory of definite formal vectors for the indefinite ones. This should
be improved on later.

Note that various issues with frontend (NB18-014, NB19-023) and flow
analysis (NB19-026) prevent enabling tests for indefinite formal vectors
for now.

* share/spark/theories/ada__containers__formal_indefinite_vectors.mlw
* ../ada__containers__formal_indefinite_vectors__generic_sorting.mlw
New theory for formal indefinite vectors.

* share/spark/theories/ada__containers__formal_vectors.mlw
* ../spark/theories/ada__containers__formal_vectors__generic_sorting.mlw
Add comments to track correspondence between definite and indefinite
versions.

* testsuite/gnatprove/tests/N317-021__vectors/indefinite_bounded.adb
* testsuite/gnatprove/tests/N317-021__vectors/indefinite_bounded.ads
* testsuite/gnatprove/tests/N317-021__vectors/indefinite_unbounded.adb
* testsuite/gnatprove/tests/N317-021__vectors/indefinite_unbounded.ads
* testsuite/gnatprove/tests/N317-021__vectors/test.py
* testsuite/gnatprove/tests/N317-021__vectors/test_vectors.adb
New tests for formal indefinite vectors, not yet enabled.