N317-021 Prepare for support of classwide elements in formal vectors
Formal indefinite vectors should be able to hold classwide elements, once
the current problems with frontend and flow analysis are fixed. This patch
fixes a minor issue which was preventing the name of formal parameters
of classwide types to be visible in the theories of formal vectors.

* gnat2why/why/gnat2why-external_axioms.adb
(Compute_Substitution_Package): Use name of underlying tagged type for
  classwide formal parameters.