N122-038 fix --RTS switch
The --RTS switch never really worked, this is now fixed. Gnatprove
itself computes (tries to) the full path to the runtime which is then
passed to gprbuild via the configuration file. The --RTS option is not
passed to gprbuild (which ignores it anyway when --config is also
present). The argument to --RTS is interpreted to be either a full path
or relative path wrt the current dir, or a path in
<spark-install>/share/spark/runtimes.

* configuration.adb:
(Set_RTS_Dir) postprocess the argument of --RTS according to description
above
* gnatproved.adb:
(Replace_Config_File_If_Needed) postprocess the
frames.cgpr/gnat2why.cgpr file to include runtime dir information