NC05-021 Only issue errors/warnings on code marked SPARK_Mode On
SPARK analysis for the generation of data and flow dependencies
should not issue errors/warnings about unsupported features on code
not marked SPARK_Mode On. This was done for most messages, but not all,
in particular not messages for unsupported pragmas. Now done.

* spark_definition.adb
Always use test on Emit_Messages with a test on the value of SPARK_Mode,
before issuing a message.

* testsuite/gnatprove/tests/NC05-021__warnings/main.adb
* testsuite/gnatprove/tests/NC05-021__warnings/test.py
* testsuite/gnatprove/tests/NC05-021__warnings/use_seq_io.adb
* testsuite/gnatprove/tests/NC05-021__warnings/use_seq_io.ads
Add test