NC05-035 Fix issue with access to computed globals before being computed
The distinction between procedures marked No_Return for the main loop and
error signaling procedures is whether the procedure has an output. If the
outputs are generated, then we should be careful when querying in which
case a procedure calls *before* the outputs are computed. In that case,
we should ignore the computed outputs, and act as if the procedure has
some output (the worst case) so that effects are properly propagated
in flow analysis during the output computation phase.

* gnat2why/flow/flow-analysis.adb,
  gnat2why/flow/flow-antialiasing.adb,
  gnat2why/flow/flow-interprocedural.adb,
  gnat2why/flow/flow_classwide.adb,
  gnat2why/spark/spark_frame_conditions.adb,
  gnat2why/why/gnat2why-subprograms.adb
Adapt to new signature of Get_Globals and Get_Proof_Globals.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Procedure_Call_Statement): Only use computed globals after they have
  been computed.

* gnat2why/flow/flow_utility.adb|s
(Get_Global, Get_Proof_Globals): Add out parameter Computed to denote that
  the globals returned were computed instead of being specified.

* gnat2why/spark/spark_util.adb|s
(Has_Output): Add parameter Use_Computed_Globals to pass to
  Get_Proof_Globals.